review: move err norm into driver, change Q to Promise and make all tests pass
